Develop

+++ Memory.h
@@ -8,9 +8,9 @@ class Memory : public MemoryInterface
 {
 private:
   uint8_t *memory;
-  uint16_t size;
+  uint32_t size;
 public:
-  Memory(uint16_t size);
+  Memory(uint32_t size);
   virtual ~Memory(void);
   virtual uint8_t ReadByte(uint16_t address) const;
   virtual void WriteByte(uint16_t address, uint8_t data);
